+++ src/command/cheque/cheque-command.ts
@@ -15,7 +15,7 @@ export class ChequeCommand extends RootCommand {
   }
 
   protected async getCashableCheques(): Promise<Cashable[]> {
-    const { lastcheques } = await this.beeDebug.getLastCheques()
+    const { lastcheques } = await this.bee.getLastCheques()
 
     const results: Cashable[] = []
     for (const cheque of lastcheques) {
@@ -41,7 +41,7 @@ export class ChequeCommand extends RootCommand {
 
   protected async getUncashedAmount(address: string): Promise<bigint> {
     try {
-      const lastCashout = await this.beeDebug.getLastCashoutAction(address)
+      const lastCashout = await this.bee.getLastCashoutAction(address)
 
       return BigInt(lastCashout.uncashedAmount)
     } catch (error) {