+++ src/command/stamp/topup.ts
@@ -26,21 +26,23 @@ export class Topup extends StampCommand implements LeafCommand {
     await super.init()
 
     if (!this.stamp) {
-      this.stamp = await pickStamp(this.beeDebug, this.console)
+      this.stamp = await pickStamp(this.bee, this.console)
     }
 
-    const spinner = createSpinner('Topup in progress. This may take a while.')
+    const spinner = createSpinner('Topup in progress. This may take a few minutes.')
 
     if (this.verbosity !== VerbosityLevel.Quiet && !this.curl) {
       spinner.start()
     }
 
     try {
-      await this.beeDebug.topUpBatch(this.stamp, this.amount.toString())
+      await this.bee.topUpBatch(this.stamp, this.amount.toString())
     } finally {
       spinner.stop()
     }
 
-    await this.printDepthAndAmount(this.stamp)
+    this.console.log(`Topup finished. Your Bee node will soon synchronize the new values from the blockchain.`)
+    this.console.log(`This can take a few minutes until the value is updated.`)
+    this.console.log(`Check it later with swarm-cli stamp show ${this.stamp}`)
   }
 }