+++ test/command/stamp.spec.ts
@@ -1,7 +1,6 @@
-import { Types } from 'cafe-utility'
+import { System, Types } from 'cafe-utility'
 import inquirer from 'inquirer'
 import { Buy } from '../../src/command/stamp/buy'
-import { sleep } from '../../src/utils'
 import { toMatchLinesInOrder } from '../custom-matcher'
 import { describeCommand, invokeTestCli } from '../utility'
 
@@ -14,38 +13,32 @@ describeCommand(
   ({ consoleMessages, getLastMessage, getNthLastMessage, hasMessageContaining }) => {
     it('should list stamps', async () => {
       await invokeTestCli(['stamp', 'list'])
-      const pattern = [['Stamp ID'], ['Usage'], ['Remaining Capacity'], ['TTL'], ['Expires']]
+      const pattern = [['Stamp ID'], ['Usage'], ['Capacity'], ['TTL']]
       expect(consoleMessages).toMatchLinesInOrder(pattern)
     })
 
     it('should show a specific stamp', async () => {
       await invokeTestCli(['stamp', 'show', Types.asString(process.env.STAMP)])
-      const pattern = [
-        ['Stamp ID', Types.asString(process.env.STAMP)],
-        ['Usage'],
-        ['Remaining Capacity'],
-        ['TTL'],
-        ['Expires'],
-      ]
+      const pattern = [['Stamp ID', Types.asString(process.env.STAMP)], ['Usage'], ['Capacity'], ['TTL']]
       expect(consoleMessages).toMatchLinesInOrder(pattern)
     })
 
     it('should not allow buying stamp with amount 0', async () => {
       await invokeTestCli(['stamp', 'buy', '--amount', '0', '--depth', '20'])
       expect(getLastMessage()).toContain('[amount] must be at least 1')
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it('should not allow buying stamp with depth 16', async () => {
       await invokeTestCli(['stamp', 'buy', '--amount', '1', '--depth', '16'])
       expect(getLastMessage()).toContain('[depth] must be at least 17')
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it('should buy stamp', async () => {
       await invokeTestCli(['stamp', 'buy', '--amount', '100000', '--depth', '20', '--yes'])
       expect(getLastMessage()).toContain('Stamp ID:')
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it('should buy stamp with immutable flag', async () => {
@@ -64,19 +57,19 @@ describeCommand(
 
       const id = command.postageBatchId
       await invokeTestCli(['stamp', 'show', id, '--verbose'])
-      const pattern = [['Total Capacity (immutable)']]
+      const pattern = [['Capacity (immutable)']]
       expect(consoleMessages).toMatchLinesInOrder(pattern)
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it.skip('should print custom message when there are no stamps', async () => {
-      await invokeTestCli(['stamp', 'list', '--bee-debug-api-url', 'http://localhost:11635'])
+      await invokeTestCli(['stamp', 'list', '--bee-api-url', 'http://localhost:11633'])
       expect(getNthLastMessage(4)).toContain('You do not have any stamps.')
     })
 
     it('should list with sorting and filter', async () => {
       await invokeTestCli(['stamp', 'list', '--min-usage', '0', '--max-usage', '100', '--least-used', '--limit', '1'])
-      const pattern = [['Stamp ID'], ['Usage'], ['Remaining Capacity'], ['TTL'], ['Expires']]
+      const pattern = [['Stamp ID'], ['Usage'], ['Capacity'], ['TTL']]
       expect(consoleMessages).toMatchLinesInOrder(pattern)
     })
 
@@ -112,7 +105,7 @@ describeCommand(
       const command = execution.runnable as Buy
       expect(command.yes).toBe(true)
       expect(inquirer.prompt).toHaveBeenCalledTimes(1)
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it('should reject estimate cost prompt', async () => {
@@ -121,7 +114,7 @@ describeCommand(
       const command = execution.runnable as Buy
       expect(command.yes).toBe(false)
       expect(inquirer.prompt).toHaveBeenCalledTimes(1)
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it('should be possible to buy with underscores and units', async () => {
@@ -139,7 +132,7 @@ describeCommand(
       const command = execution.runnable as Buy
       expect(command.yes).toBe(true)
       expect(getLastMessage()).toContain('Stamp ID:')
-      await sleep(11_000)
+      await System.sleepMillis(11_000)
     })
 
     it.skip('should only be able to dilute stamp with greater depth', async () => {
@@ -166,8 +159,7 @@ describeCommand(
       expect(hasMessageContaining('This postage stamp already has depth 19. The new value must be higher.')).toBe(true)
       consoleMessages.length = 0
       await invokeTestCli(['stamp', 'dilute', '--stamp', postageBatchId, '--depth', '20'])
-      expect(getNthLastMessage(2)).toContain('Depth')
-      expect(getNthLastMessage(2)).toContain('20')
+      expect(getNthLastMessage(3)).toContain('Dilute finished')
     })
 
     it.skip('should top up stamp', async () => {
@@ -188,8 +180,7 @@ describeCommand(
       const { postageBatchId } = command
       consoleMessages.length = 0
       await invokeTestCli(['stamp', 'topup', '--stamp', postageBatchId, '--amount', '1k'])
-      expect(getNthLastMessage(1)).toContain('Amount')
-      expect(getNthLastMessage(1)).toContain('2000')
+      expect(getNthLastMessage(3)).toContain('Topup finished')
     })
   },
 )